move test_cellmodule_contents to collection tests
